change where autobone video is
